+++ R/aaa.R
@@ -90,7 +90,8 @@ utils::globalVariables(
     "neighbors", ".submodels", "has_submodel", "max_neighbor", "max_penalty",
     "max_terms", "max_tree", "model", "name", "num_terms", "penalty", "trees",
     "sub_neighbors", ".pred_class", "x", "y", "predictor_indicators",
-    "compute_intercept", "remove_intercept", "estimate", "term")
+    "compute_intercept", "remove_intercept", "estimate", "term",
+    "call_info", "component", "component_id", "func")
 )
 
 # nocov end