+++ R/predict_quantile.R
@@ -1,13 +1,17 @@
 #' @keywords internal
 #' @rdname other_predict
-#' @param quant A vector of numbers between 0 and 1 for the quantile being
+#' @param quantile A vector of numbers between 0 and 1 for the quantile being
 #'  predicted.
 #' @inheritParams predict.model_fit
 #' @method predict_quantile model_fit
 #' @export predict_quantile.model_fit
 #' @export
-predict_quantile.model_fit <-
-  function(object, new_data, quantile = (1:9)/10, ...) {
+predict_quantile.model_fit <- function(object,
+                                       new_data,
+                                       quantile = (1:9)/10,
+                                       interval = "none",
+                                       level = 0.95,
+                                       ...) {
 
     check_spec_pred_type(object, "quantile")