+++ R/surv_reg_flexsurv.R
@@ -1,0 +1,9 @@
+#' Parametric survival regression
+#'
+#' [flexsurv::flexsurvreg()] fits a parametric survival model.
+#'
+#' @includeRmd man/rmd/surv_reg_flexsurv.Rmd details
+#'
+#' @name details_surv_reg_flexsurv
+#' @keywords internal
+NULL