+++ man-roxygen/spec-survival.R
@@ -1,0 +1,4 @@
+#' @details
+#' Since survival models typically involve censoring (and require the use of
+#' [survival::Surv()] objects), the [fit.model_spec()] function will require that the
+#' survival model be specified via the formula interface.