+++ NEWS.md
@@ -10,6 +10,18 @@
 
 * Fixed a bug for `logistic_reg()` with the LiblineaR engine (#552).
 
+* The list column produced when creating survival probability predictions is now always called `.pred` (with `.pred_survival` being used inside of the list column). 
+
+## Other Changes
+
+* When the xy interface is used and the underlying model expects to use a matrix, a better warning is issued when predictors contain non-numeric columns (including dates). 
+ 
+* The fit time is only calculated when the `verbosity` argument of `control_parsnip()` is 2L or greater. Also, the call to `system.time()` now uses `gcFirst = FALSE`. (#611)
+ 
+* `fit_control()` is soft-deprecated in favor of `control_parsnip()`. 
+
+* Argument `interval` was added for prediction: For types "survival" and "quantile", estimates for the confidence or prediction interval can be added if available (#615).
+
 
 # parsnip 0.1.7